+++ app/[locale]/projects/[projectId]/collaboration/actions/index.tsx
@@ -1,0 +1,3 @@
+export { updateProjectCollaboratorAction } from './updateProjectCollaboratorAction';
+export { removeProjectCollaboratorAction } from './removeProjectCollaboratorAction';
+export { resendInviteToProjectCollaboratorAction } from './resendInviteToProjectCollaborator';