+++ utils/errorHandler.ts
@@ -94,19 +94,27 @@ export const handleErrors = async (
 /**
  * Shared function to extract specified errors from an error object.
  * @param errs - The error object (can be any shape)
- * @param keys - The keys to extract errors from
+ * @param keys - (Optional)- The keys to extract errors from
  * @returns Array of error messages (non-empty strings)
  */
 export function extractErrors<T extends Record<string, string | undefined>>(
   errs: T,
-  keys: (keyof T)[]
+  keys?: (keyof T)[]
 ): string[] {
   const newErrors: string[] = [];
 
-  for (const key of keys) {
-    const val = errs[key];
-    if (val) {
-      newErrors.push(val);
+  if (keys && keys.length > 0) {
+    for (const key of keys) {
+      const val = errs[key];
+      if (val) {
+        newErrors.push(val);
+      }
+    }
+  } else {
+    for (const val of Object.values(errs)) {
+      if (val) {
+        newErrors.push(val);
+      }
     }
   }